+++ cvat/apps/webhooks/event_type.py
@@ -5,7 +5,7 @@
 from .models import WebhookTypeChoice
 
 
-def event_name(action, resource):
+def event_name(action: str, resource: str) -> str:
     return f"{action}:{resource}"
 
 
@@ -24,7 +24,7 @@ class Events:
     @classmethod
     def select(cls, resources):
         return [
-            f"{event_name(action, resource)}"
+            event_name(action=action, resource=resource)
             for resource in resources
             for action in cls.RESOURCES.get(resource, [])
         ]
@@ -39,7 +39,7 @@ class EventTypeChoice:
 class AllEvents:
     webhook_type = "all"
     events = list(
-        event_name(action, resource)
+        event_name(action=action, resource=resource)
         for resource, actions in Events.RESOURCES.items()
         for action in actions
     )
@@ -49,8 +49,8 @@ class ProjectEvents:
     webhook_type = WebhookTypeChoice.PROJECT
     events = [
         *Events.select(["task", "job", "label", "issue", "comment"]),
-        event_name("update", "project"),
-        event_name("delete", "project"),
+        event_name(action="update", resource="project"),
+        event_name(action="delete", resource="project"),
     ]
