+++ cvat/apps/webhooks/utils.py
@@ -1,0 +1,14 @@
+# Copyright (C) CVAT.ai Corporation
+#
+# SPDX-License-Identifier: MIT
+
+from cvat.apps.engine.serializers import BasicUserSerializer
+from cvat.apps.events.handlers import get_request, get_user
+
+
+def get_sender(instance) -> dict:
+    user = get_user(instance)
+    if isinstance(user, dict):
+        return user
+
+    return BasicUserSerializer(user, context={"request": get_request(instance)}).data
